+++ .dotnet/src/Custom/Files/FileClient.cs
@@ -51,6 +51,7 @@ public partial class FileClient
 
         _pipeline = OpenAIClient.CreatePipeline(credential, options);
         _endpoint = OpenAIClient.GetEndpoint(options);
+        _internalUploadsClient = new(_pipeline, options);
     }
 
     // CUSTOM: